Crate domain_service

Service and ServiceFactory implementation. Specialized wrapper over substrate service.

Modules

config

providers

rpc

A collection of node-specific RPC methods. Substrate provides the sc-rpc crate, which defines the core RPC layer used by Substrate nodes. This file extends those RPC definitions with capabilities that are specific to this project’s runtime configuration.

Structs

DomainParams

NewFull

Domain full node along with some other components.

Functions

build_network

Build the network service, the network status sinks and an RPC sender.

new_full

Builds service for a domain full node.

Type Aliases

DomainOperator

FullBackend

FullClient

Domain full client.

FullPool