Function domain_service::new_full

pub async fn new_full<PBlock, PClient, SC, IBNS, NSNS, RuntimeApi, ExecutorDispatch>(
    secondary_chain_config: Configuration,
    primary_chain_client: Arc<PClient>,
    primary_network: Arc<NetworkService<PBlock, PBlock::Hash>>,
    select_chain: &SC,
    imported_block_notification_stream: IBNS,
    new_slot_notification_stream: NSNS,
    block_import_throttling_buffer_size: u32
) -> Result<NewFull<Arc<FullClient<RuntimeApi, ExecutorDispatch>>, NativeElseWasmExecutor<ExecutorDispatch>, PBlock, PClient, RuntimeApi, ExecutorDispatch>>where
    PBlock: BlockT,
    PClient: HeaderBackend<PBlock> + BlockBackend<PBlock> + ProvideRuntimeApi<PBlock> + Send + Sync + 'static,
    PClient::Api: ExecutorApi<PBlock, Hash>,
    SC: SelectChain<PBlock>,
    IBNS: Stream<Item = (NumberFor<PBlock>, ForkChoiceStrategy, Sender<()>)> + Send + 'static,
    NSNS: Stream<Item = (Slot, Blake2b256Hash)> + Send + 'static,
    RuntimeApi: ConstructRuntimeApi<Block, FullClient<RuntimeApi, ExecutorDispatch>> + Send + Sync + 'static,
    RuntimeApi::RuntimeApi: ApiExt<Block, StateBackend = StateBackendFor<TFullBackend<Block>, Block>> + Metadata<Block> + BlockBuilder<Block> + OffchainWorkerApi<Block> + SessionKeys<Block> + DomainCoreApi<Block, AccountId> + SystemDomainApi<Block, NumberFor<PBlock>, PBlock::Hash> + TaggedTransactionQueue<Block> + AccountNonceApi<Block, AccountId, Nonce> + TransactionPaymentRuntimeApi<Block, Balance> + RelayerApi<Block, RelayerId, NumberFor<Block>>,
    ExecutorDispatch: NativeExecutionDispatch + 'static,

Start a node with the given parachain Configuration and relay chain Configuration.

This is the actual implementation that is abstract over the executor and the runtime api.