Function domain_service::new_full_system

pub async fn new_full_system<PBlock, PClient, SC, IBNS, CIBNS, NSNS, RuntimeApi, ExecutorDispatch>(
    system_domain_config: DomainConfiguration,
    primary_chain_client: Arc<PClient>,
    primary_network_sync_oracle: Arc<dyn SyncOracle + Send + Sync>,
    select_chain: &SC,
    executor_streams: ExecutorStreams<PBlock, IBNS, CIBNS, NSNS>,
    gossip_message_sink: GossipMessageSink
) -> Result<NewFullSystem<Arc<FullClient<RuntimeApi, ExecutorDispatch>>, NativeElseWasmExecutor<ExecutorDispatch>, PBlock, PClient, RuntimeApi, ExecutorDispatch>>where
    PBlock: BlockT,
    NumberFor<PBlock>: From<NumberFor<Block>>,
    PBlock::Hash: From<Hash>,
    PClient: HeaderBackend<PBlock> + HeaderMetadata<PBlock, Error = Error> + BlockBackend<PBlock> + ProvideRuntimeApi<PBlock> + BlockchainEvents<PBlock> + Send + Sync + 'static,
    PClient::Api: ExecutorApi<PBlock, Hash>,
    SC: SelectChain<PBlock>,
    IBNS: Stream<Item = (NumberFor<PBlock>, Sender<()>)> + Send + 'static,
    CIBNS: Stream<Item = BlockImportNotification<PBlock>> + Send + 'static,
    NSNS: Stream<Item = (Slot, Blake2b256Hash, Option<Sender<()>>)> + Send + 'static,
    RuntimeApi: ConstructRuntimeApi<Block, FullClient<RuntimeApi, ExecutorDispatch>> + Send + Sync + 'static,
    RuntimeApi::RuntimeApi: ApiExt<Block, StateBackend = StateBackendFor<TFullBackend<Block>, Block>> + Metadata<Block> + BlockBuilder<Block> + OffchainWorkerApi<Block> + SessionKeys<Block> + DomainCoreApi<Block, AccountId> + SystemDomainApi<Block, NumberFor<PBlock>, PBlock::Hash> + MessengerApi<Block, NumberFor<Block>> + TaggedTransactionQueue<Block> + AccountNonceApi<Block, AccountId, Nonce> + TransactionPaymentRuntimeApi<Block, Balance> + RelayerApi<Block, RelayerId, NumberFor<Block>> + ReceiptsApi<Block, Hash> + PreValidationObjectApi<Block, Hash>,
    ExecutorDispatch: NativeExecutionDispatch + 'static,

Start a node with the given system domain Configuration and consensus chain Configuration.

This is the actual implementation that is abstract over the executor and the runtime api.